feat: remove unnecessary audio extraction
Turnsout the audio extraction before sending the file to Whisper does not save much time, if any. Removed to reduce complexity.

feat: identify audio language before sending the file to Whisper

Noticed this is absolutely necessary for videos not containing any spoken language in the first 30 seconds. If this is not set the returned subtitle will be a weird mess.

feat: add checks to automatically review the generated subtitle files for error responses

If whisper-asr-webservice responds with an error, curl will write it into the SRT file as content. The script now checks the content for 'Internal Server Error' or a '{' at the start (indicating a JSON response instead of SRT). In both cases the error is logged and the generated file deleted.